@hydroo 2023-08-26
I'll dig a bit into the other quants and QK_K != 256.
In the meantime I'm interested in any feedback.
Thanks!