@JackJollimore 2023-08-26
Hi , I'm having an issue loading from this PR. 

```
./main -m ~/llama-2-chat.gguf
```

I converted llama-2-7b-chat.ggmlv3.q4_0.bin to `gguf`, so I expect it to load. Here's the error:
```
main: build = 930 (a7e10cb)                                
main: seed  = 1693075923                                   
llama.cpp: loading model from /data/data/com.termux/files/home/zarafusionex-1.1-l2-7b.Q4_K_M.gguf                     
error loading model: unknown (magic, version) combination: 46554747, 00000001; is this really a GGML file?            
llama_load_model_from_file: failed to load model           
llama_init_from_gpt_params: error: failed to load model '/data/data/com.termux/files/home/zarafusionex-1.1-l2-7b.Q4_K_M.gguf'                                                    
main: error: unable to load model
```
Same error with zarafusionex-1.1-l2-7b.Q4_K_M.gguf. 

These models load as expected on master `llama.cpp`. Maybe it's obvious issue, but I don't see it.