@slaren 2023-08-26
You need to merge master first before testing, the source branch is very outdated. However, this PR only affects x86 processors with AVX2, you won't be able to test these changes on an Android ARM device.