@hydroo 2023-08-26
This is partially my bad because I initiated the change before the gguf merge, and wanted to try it out.
I'll try changing my workflow to gguf now.