@slaren 2023-08-26
Quick test with 13900K/24 threads/7B:

Master:
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | pp 64      |     58.37 ± 0.14 |
| LLaMA 7B mostly Q4_K - Medium  |   3.80 GiB |     6.74 B | CPU        |         24 | pp 64      |     55.96 ± 0.20 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | tg 32      |     18.02 ± 0.18 |
| LLaMA 7B mostly Q4_K - Medium  |   3.80 GiB |     6.74 B | CPU        |         24 | tg 32      |     17.32 ± 0.06 |

build: c7d92e6 (1081)


PR:
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | pp 64      |     57.46 ± 0.13 |
| LLaMA 7B mostly Q4_K - Medium  |   3.80 GiB |     6.74 B | CPU        |         24 | pp 64      |     55.20 ± 0.18 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | tg 32      |     18.15 ± 0.13 |
| LLaMA 7B mostly Q4_K - Medium  |   3.80 GiB |     6.74 B | CPU        |         24 | tg 32      |     17.27 ± 0.07 |

build: ff83017 (1082)

Not sure if I am doing something wrong, maybe the difference is bigger with 13B?