@slaren 2023-08-26
Similar results with 13B:

| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 13B mostly Q4_K - Small  |   6.90 GiB |    13.02 B | CPU        |         24 | pp 64      |     30.77 ± 0.02 |
| LLaMA 13B mostly Q4_K - Small  |   6.90 GiB |    13.02 B | CPU        |         24 | tg 24      |      9.39 ± 0.04 |

build: c7d92e6 (1081)

| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 13B mostly Q4_K - Small  |   6.90 GiB |    13.02 B | CPU        |         24 | pp 64      |     30.80 ± 0.05 |
| LLaMA 13B mostly Q4_K - Small  |   6.90 GiB |    13.02 B | CPU        |         24 | tg 24      |      9.50 ± 0.10 |

build: ff83017 (1082)

Compiler: `g++ (Ubuntu 12.3.0-1ubuntu1~23.04) 12.3.0`