@hydroo 2023-08-26
I'm doing more tests now. Would you mind sharing your command?
I'm using MSVC 2022 17.5.2 .