@slaren 2023-08-26
These are the commands I used:
```bash
./llama-bench -m models/7B/ggml-model-Q4_K_S.gguf -m models/7B/ggml-model-Q4_K_M.gguf -n 32 -p 64 -t 24 
./llama-bench -m models/13B/ggml-model-Q4_K_S.gguf -n 24 -p 64 -t 24    
```