@hydroo 2023-08-26
Where is this llama-bench from?

I'm still getting pretty reliable speedups as quotes above.

For 7b q4_K_S I get 11.9-12.1t/s vs 13.3-13.6 with the change