@slaren 2023-08-26
`llama-bench` is an example in this repository. If you used cmake to build, you should have the executable.

The difference may be easily explained by the different compilers, MSVC is notoriously slower than GCC these days.