@slaren 2023-08-27
Results under native Windows with MSVC (please ignore the broken utf-8):

| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | pp 64      |     51.41 ┬▒ 0.19 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | tg 32      |     16.64 ┬▒ 0.03 |

build: 25423e9 (1090)



| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | pp 64      |     51.62 ┬▒ 0.11 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         24 | tg 32      |     16.22 ┬▒ 0.07 |

build: ff83017 (1082)


Compiler: `Microsoft (R) C/C++ Optimizing Compiler Version 19.29.30151 for x64`