@ikawrakow 2023-08-27
I wrote the code that is being optimized in this PR. I wrote it the way it is after experimenting what is the fastest way to do this calculation (including testing the version that is being proposed here). I was testing on Ryzen 7950X, Ryzen 5950X, and on i9-9960X. The version that landed in the code was tiny little bit faster than what is being proposed (and still is on these 3 computers). E.g., on Ryzen 7950X:

Master:
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         16 | pp 64      |     84.48 ± 1.11 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         16 | tg 32      |     14.86 ± 0.02 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |          8 | pp 64      |     46.78 ± 0.38 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |          8 | tg 32      |     15.14 ± 0.01 |

PR:
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         16 | pp 64      |    73.44 ± 15.16 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |         16 | tg 32      |     14.76 ± 0.02 
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |          8 | pp 64      |     45.78 ± 0.40 |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB |     6.74 B | CPU        |          8 | tg 32      |     15.08 ± 0.13 |