@hydroo 2023-08-28
Running with -t24 and boost maybe introduces effects that we don't want to measure.
I'm doing the following benchmarks with clock-locking (no boost) and -t 8 to stay on P-cores.
(And every measurement twice, and merged the tables)

**codellama-13b-instruct.Q4_K_S.gguf**
Before:
llama-bench -m codellama-13b-instruct.Q4_K_S.gguf -n 24 -p 64 -t 8
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | pp 64      |     10.09 ± 0.31 |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | pp 64      |     11.41 ± 0.29 |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | tg 24      |      5.12 ± 0.33 | 
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | tg 24      |      5.41 ± 0.33 |

After:
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | pp 64      |     10.63 ± 0.53 |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | pp 64      |     12.29 ± 0.99 |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | tg 24      |      5.61 ± 0.16 |
| LLaMA 13B mostly Q4_K - Small  |   7.08 GiB |    13.02 B | CPU        |          8 | tg 24      |      5.40 ± 0.29 |

The PP has ~20% variance. More than the proposed optimization.
The TG has ~7% - borderline.
Overall, for this model, the PR doesn't look beneficial.

**codellama-34b-instruct.Q4_K_S.gguf**
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | pp 64      |      5.51 ± 0.20 |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | pp 64      |      5.61 ± 0.36 |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | tg 24      |      2.18 ± 0.08 |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | tg 24      |      2.09 ± 0.10 |

After:
| model                          |       size |     params | backend    |    threads | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ---------- | ---------------: |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | pp 64      |      5.19 ± 0.22 |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | pp 64      |      5.78 ± 0.24 |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | tg 24      |      1.98 ± 0.02 |
| LLaMA 34B mostly Q4_K - Small  |  17.83 GiB |    33.74 B | CPU        |          8 | tg 24      |      2.02 ± 0.04 |

PP is again unreliable.
TG is better by ~5-10%.

In VTune I observe ~0.5% fewer instructions retired in the function.
The code gen is a hard to grok.
I could immediately make out major differences.

@ikawrakow would you mind trying a larger model?
Or do you think it's too slow, i.e. not a use-case and therefore we shouldn't compare those?